\name{multicore-package}
\alias{multicore-package}
\alias{multicore}
\alias{SIGCHLD}
\alias{SIGHUP}
\alias{SIGINT}
\alias{SIGKILL}
\alias{SIGQUIT}
\alias{SIGSTOP}
\alias{SIGTERM}
\alias{SIGUSR1}
\alias{SIGUSR2}
\alias{children}
\alias{collect}
\alias{exit}
\alias{fork}
\alias{kill}
\alias{mclapply}
\alias{mcparallel}
\alias{parallel}
\alias{processID}
\alias{pvec}

\title{
The 'multicore' Package
}
\description{
  Package \pkg{multicore} was superseded by \pkg{parallel} in 2011.  This
  is a transitional stub which will be removed during 2014.

  Most of the objects previously provided by \pkg{multicore} have been
  imported here from \pkg{parallel} or \pkg{tools} and exported.

  Objects \code{SIGALARM}, \code{SIGINFO}, \code{readChild},
  \code{readChildren}, \code{selectChildren}, \code{sendChildStdin} and
  \code{sendMaster} are no longer available.
}
\examples{
### consult the corresponding help pages in package 'parallel'
}
\keyword{ package }
